+++ lib/config/router/router.dart
@@ -1,4 +1,5 @@
-import 'package:base_project/presentation/widgets/widget_tree.dart';
+import 'package:base_project/presentation/widget_tree.dart';
+import 'package:base_project/screens/calendar/screenParaFerchi_BB.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import '../../screens/index.dart';
@@ -15,13 +16,34 @@ final GoRouter router = GoRouter(
     GoRoute(
       path: '/login',
       builder: (BuildContext context, GoRouterState state) {
-        return const LoginPage();
+        return const LoginScreen();
       },
     ),
     GoRoute(
-      path: '/test',
+      path: '/calendar',
       builder: (BuildContext context, GoRouterState state) {
-        return Test();
+        return CalendarScreen();
+      },
+    ),
+    GoRoute(
+      path: '/recipes',
+      builder: (BuildContext context, GoRouterState state) {
+        return const RecipesScreen();
+      },
+    ),
+    GoRoute(
+      path: '/recipes/clendar/:date1/:date2/:recipeParam',
+      builder: (BuildContext context, GoRouterState state) {
+        return CalenderRecipe(
+            date1: state.pathParameters['date1']!,
+            date2: state.pathParameters['date2']!,
+            recipeParam: state.pathParameters['recipeParam']);
+      },
+    ),
+    GoRoute(
+      path: '/recipes/:id',
+      builder: (BuildContext context, GoRouterState state) {
+        return EditRecipeScreen(id: int.parse(state.pathParameters['id']!));
       },
     ),
   ],