+++ lib/models/recipe/recipe_model.dart
@@ -1,0 +1,17 @@
+class RecipeModel {
+  final int id;
+  final String name;
+  final String description;
+
+  RecipeModel({
+    required this.id, 
+    required this.name, 
+    required this.description
+  });
+
+  factory RecipeModel.fromJson(Map<String, dynamic> json) => RecipeModel(
+    id: int.parse(json['id']), 
+    name: json['name'], 
+    description: json['description']
+  );
+}