+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This file is used to explain in detail changes made to the food structured app.
 <!-- TOC -->
   [[TOC]]
 
+## V 1.1.0
+Date: Nov 26, 2023
+* [NEW]
+  * Calendar screen
+  * Recipes screen and edit recipe screen
+  * Sidebar drawer
+* [UPDATE]
+  * Logout bugfix
+
 ## V 1.0.0
 Date: Nov 05, 2023
 * [NEW]